Add files via upload
The LED is mounted on arduino which is connected to raspberry pi.
Using raspberry pi we get the feature of wifi to arduino. 
The LED can be controlled using the switch button, designed with one of the templates of “io.adafruit.com”.
This implies the concept of the Internet of Things.
Hence the LED can be controlled by the button from anywhere at anytime.